requirements.txt
boto3
python-dotenv
wget
requests
beautifulsoup4
psycopg2-binary